+++ envconfig/envconfig.go
@@ -1,4 +1,4 @@
-package config
+package envconfig
 
 import (
 	"bytes"
